JAMES-1854 Move Sieve related classes in SIEVE package
